WIP: integrate react-polymorph package
