+++ app/components/forms/CurrencyInputSkin.scss
