--- app/components/forms/TextInputSkin.scss
@@ -1,55 +1,0 @@
-.component {
-  position: relative;
-}
-
-.label {
-  font-size: 16px;
-  font-family: $font-medium;
-  color: #5e6066;
-  display: block;
-  margin-bottom: 10px;
-}
-
-.error {
-  font-family: $font-regular;
-  font-size: 14px;
-  color: #ea4c5b;
-  opacity: 0.5;
-  position: absolute;
-  top: 0;
-  right: 0;
-}
-
-.input {
-  width: 100%;
-  font-size: 16px;
-  font-family: $font-light;
-  border: 1px solid #5e6066;
-  padding: 14px 20px;
-  color: #5e6066;
-
-  &:focus {
-    outline: none;
-  }
-
-  &::-webkit-input-placeholder {
-    color: #5e6066;
-    opacity: 0.5;
-  }
-}
-
-.disabled {
-  .label {
-    color: #b6b6b6;
-  }
-  .input {
-    border-color: #cfcfcf;
-    background-color: #fcfafa;
-  }
-}
-
-.errored {
-  .input {
-    border-color: #ea4c5b;
-  }
-}
