+++ app/themes/daedalus.js
@@ -1,3 +1,4 @@
+import simpleTheme from 'react-polymorph/lib/themes/simple/theme';
 import overlay from 'react-toolbox/lib/overlay/theme.scss';
 import navigation from 'react-toolbox/lib/navigation/theme.scss';
 import dropdown from './daedalus/dropdown.scss';
@@ -7,6 +8,18 @@ import button from './daedalus/button.scss';
 import appBar from './daedalus/app-bar.scss';
 import switchStyle from './daedalus/switch.scss';
 import checkbox from './daedalus/checkbox.scss';
+
+// export const daedalusTheme = Object.assign({}, simpleTheme, {
+//   RTInput: input,
+//   RTDropdown: dropdown,
+//   RTOverlay: overlay,
+//   RTNavigation: navigation,
+//   RTDialog: dialog,
+//   RTButton: button,
+//   RTAppBar: appBar,
+//   RTSwitch: switchStyle,
+//   RTCheckbox: checkbox
+// });
 
 export const daedalusTheme = {
   RTInput: input,
